src/requirements.txt
numpy==1.22.4
Flask==2.2.2
tflite-runtime==2.10.0
Pillow==9.2.0
joblib==1.3.2
pandas==2.1.1
scikit-learn==1.3.1